Introduce assert_equals with dynamic test naming
We started by writing a main method with an if statement checking the
output of fizzbuzz(1) for correctness and print the results of the checks.
We then extracted the duplicated constants (expected and actual values)
into variables. Next we added a new test, fizzbuzz(2) == '2', by copying
and pasting the first test, we then extracted the common logic of checking
the output of the function against the expected value to an `assert_equals`
method which took the test name. We added another two tests for
`fizzbuzz(3) == 'Fizz'` and `fizzbuzz(6) == 'Fizz'`. Finally we replaced
extracted each call to `assert_equals` to its own method with the testname
embedded in the method name. To reduce duplication of the test name we
then got the calling test method's name by inspecting the stack.